requirements.txt
talisker[gunicorn]==0.14.3
canonicalwebteam.http==1.0.1
canonicalwebteam.yaml-responses[flask]==1.1.1
flask==1.0.2
feedparser==5.2.1